Service to add new Doctor and Appointment, and samll changes

+++ src/main/java/com/vetclinic/demo/model/Service.java
@@ -10,7 +10,7 @@ public class Service {
     private Long id;
 
     private String name;
-    private Long price;
+    private double price;
 
 
     public Service(String name, Long price) {
@@ -20,6 +20,7 @@ public class Service {
 
     public Service() {
     }
+
 
     public Long getId() {
         return id;
@@ -37,7 +38,7 @@ public class Service {
         this.name = name;
     }
 
-    public Long getPrice() {
+    public double getPrice() {
         return price;
     }
