+++ src/main/java/com/vetclinic/demo/model/request/ServiceRequest.java
@@ -2,7 +2,8 @@ package com.vetclinic.demo.model.request;
 
 public class ServiceRequest {
     private String name;
-    private Long price;
+    private double price;
+
 
     public String getName() {
         return name;
@@ -12,7 +13,7 @@ public class ServiceRequest {
         this.name = name;
     }
 
-    public Long getPrice() {
+    public double getPrice() {
         return price;
     }
